flask_migration: ensure same features flask/webpy
